+++ backend/Lumino.API/Application/Services/AdminTopicService.cs
@@ -31,6 +31,11 @@ namespace Lumino.Api.Application.Services
 
         public AdminTopicResponse Create(CreateTopicRequest request)
         {
+            if (request == null)
+            {
+                throw new ArgumentException("Request is required");
+            }
+
             var topic = new Topic
             {
                 CourseId = request.CourseId,
@@ -52,7 +57,17 @@ namespace Lumino.Api.Application.Services
 
         public void Update(int id, UpdateTopicRequest request)
         {
-            var topic = _dbContext.Topics.First(x => x.Id == id);
+            if (request == null)
+            {
+                throw new ArgumentException("Request is required");
+            }
+
+            var topic = _dbContext.Topics.FirstOrDefault(x => x.Id == id);
+
+            if (topic == null)
+            {
+                throw new KeyNotFoundException("Topic not found");
+            }
 
             topic.Title = request.Title;
             topic.Order = request.Order;
@@ -62,7 +77,13 @@ namespace Lumino.Api.Application.Services
 
         public void Delete(int id)
         {
-            var topic = _dbContext.Topics.First(x => x.Id == id);
+            var topic = _dbContext.Topics.FirstOrDefault(x => x.Id == id);
+
+            if (topic == null)
+            {
+                throw new KeyNotFoundException("Topic not found");
+            }
+
             _dbContext.Topics.Remove(topic);
             _dbContext.SaveChanges();
         }
